Pulling example App into a separate package
